+++ 1.20/Common/src/main/java/fuzs/sneakycurses/client/SneakyCursesClient.java
@@ -30,7 +30,7 @@ public class SneakyCursesClient implements ClientModConstructor {
         });
         ClientEntityLevelEvents.LOAD.register((entity, level) -> {
             if (entity instanceof ThrownTrident) {
-                // need to go this way around as during the server load event the entity has not yet been sycned to clients
+                // need to go this way around as during the server load event the entity has not yet been synced to clients
                 SneakyCurses.NETWORK.sendToServer(new ServerboundRequestTridentItemMessage(entity.getId()));
             }
             return EventResult.PASS;