+++ 1.20/Common/src/main/java/fuzs/sneakycurses/client/handler/ItemTooltipHandler.java
@@ -3,23 +3,23 @@ package fuzs.sneakycurses.client.handler;
 import fuzs.puzzleslib.api.client.event.v1.ScreenEvents;
 import fuzs.puzzleslib.api.client.screen.v2.ScreenHelper;
 import fuzs.sneakycurses.SneakyCurses;
+import fuzs.sneakycurses.client.util.ComponentHelper;
 import fuzs.sneakycurses.config.ServerConfig;
 import fuzs.sneakycurses.handler.CurseRevealHandler;
 import net.minecraft.ChatFormatting;
-import net.minecraft.Util;
 import net.minecraft.client.Minecraft;
+import net.minecraft.client.gui.Font;
 import net.minecraft.client.gui.components.AbstractWidget;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.client.gui.screens.inventory.AnvilScreen;
 import net.minecraft.client.gui.screens.inventory.CreativeModeInventoryScreen;
+import net.minecraft.client.gui.screens.inventory.EnchantmentNames;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.network.chat.Component;
-import net.minecraft.network.chat.Style;
+import net.minecraft.network.chat.FormattedText;
 import net.minecraft.network.chat.contents.TranslatableContents;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.util.RandomSource;
 import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.inventory.AnvilMenu;
 import net.minecraft.world.inventory.Slot;
 import net.minecraft.world.item.EnchantedBookItem;
 import net.minecraft.world.item.ItemStack;
@@ -29,11 +29,10 @@ import org.jetbrains.annotations.Nullable;
 
 import java.util.List;
 import java.util.ListIterator;
+import java.util.Random;
 
 public class ItemTooltipHandler {
-    private static final RandomSource RANDOM_SOURCE = RandomSource.create();
-    private static final Style ALT_FONT_STYLE = Style.EMPTY.withFont(new ResourceLocation("alt"));
-    private static final String[] ENCHANTMENT_LORE = new String[]{"the", "elder", "scrolls", "klaatu", "berata", "niktu", "xyzzy", "bless", "curse", "light", "darkness", "fire", "air", "earth", "water", "hot", "dry", "cold", "wet", "ignite", "snuff", "embiggen", "twist", "shorten", "stretch", "fiddle", "destroy", "imbue", "galvanize", "enchant", "free", "limited", "range", "of", "towards", "inside", "sphere", "cube", "self", "other", "ball", "mental", "physical", "grow", "shrink", "demon", "elemental", "spirit", "animal", "creature", "beast", "humanoid", "undead", "fresh", "stale", "phnglui", "mglwnafh", "cthulhu", "rlyeh", "wgahnagl", "fhtagn", "baguette"};
+    private static final Random RANDOM = new Random();
 
     private static int currentScreenSeed;
 
@@ -41,9 +40,9 @@ public class ItemTooltipHandler {
         setFreshSeed();
     }
 
-    public static void onItemTooltip(ItemStack stack, @Nullable Player player, List<Component> lines, TooltipFlag context) {
-        if (!SneakyCurses.CONFIG.get(ServerConfig.class).obfuscateCurses) return;
-        if (!isAffected(player, stack)) return;
+    public static void onItemTooltip(ItemStack itemStack, @Nullable Player player, List<Component> lines, TooltipFlag context) {
+        if (!SneakyCurses.CONFIG.getHolder(ServerConfig.class).isAvailable() || !SneakyCurses.CONFIG.get(ServerConfig.class).obfuscateCurses) return;
+        if (!isAffected(player, itemStack)) return;
         ListIterator<Component> iterator = lines.listIterator();
         while (iterator.hasNext()) {
             if (iterator.next().getContents() instanceof TranslatableContents contents && contents.getKey().startsWith("enchantment.")) {
@@ -58,7 +57,7 @@ public class ItemTooltipHandler {
                 if (enchantment != null && enchantment.isCurse()) {
                     if (enchantmentKey.length == 3) {
                         initSeed(currentScreenSeed + BuiltInRegistries.ENCHANTMENT.getId(enchantment));
-                        Component component = getLoreForWidth();
+                        Component component = getLoreForWidth(Minecraft.getInstance().font);
                         iterator.set(Component.empty().append(component).withStyle(ChatFormatting.RED));
                     } else {
                         // remove descriptions from the enchantment descriptions mod, this matches their format
@@ -69,25 +68,19 @@ public class ItemTooltipHandler {
         }
     }
 
-    private static Component getLoreForWidth() {
-        int maxWidth = RANDOM_SOURCE.nextInt(16, 24);
-        StringBuilder builder = new StringBuilder();
-        while (builder.length() <= maxWidth) {
-            if (!builder.isEmpty()) builder.append(" ");
-            builder.append(Util.getRandom(ENCHANTMENT_LORE, RANDOM_SOURCE));
-        }
-        if (builder.lastIndexOf(" ") != -1) {
-            builder.delete(builder.lastIndexOf(" "), builder.length());
-        }
-        return Component.literal(builder.toString()).withStyle(ALT_FONT_STYLE);
+    private static Component getLoreForWidth(Font font) {
+        double maxWidth = RANDOM.nextGaussian(100, 20);
+        FormattedText formattedText = EnchantmentNames.getInstance().getRandomName(font, (int) maxWidth);
+        return ComponentHelper.toComponent(formattedText);
     }
 
     private static void setFreshSeed() {
-        currentScreenSeed = RANDOM_SOURCE.nextInt();
+        currentScreenSeed = RANDOM.nextInt();
     }
 
     private static void initSeed(long seed) {
-        RANDOM_SOURCE.setSeed(seed);
+        RANDOM.setSeed(seed);
+        EnchantmentNames.getInstance().initSeed(seed);
     }
 
     private static boolean isAffected(@Nullable Player player, ItemStack itemStack) {
@@ -114,7 +107,8 @@ public class ItemTooltipHandler {
                     }
                 }
             }
+            return CurseRevealHandler.isAffected(itemStack) && !CurseRevealHandler.allCursesRevealed(itemStack);
         }
-        return CurseRevealHandler.isAffected(itemStack) && !CurseRevealHandler.allCursesRevealed(itemStack);
+        return false;
     }
 }