+++ 1.20/Common/src/main/java/fuzs/sneakycurses/config/ServerConfig.java
@@ -4,15 +4,15 @@ import fuzs.puzzleslib.api.config.v3.Config;
 import fuzs.puzzleslib.api.config.v3.ConfigCore;
 
 public class ServerConfig implements ConfigCore {
-    @Config(description = "Obfuscate curse enchantments with runes on item tooltips.")
+    @Config(name = "obfuscate_curses_on_tooltips", description = "Obfuscate curse enchantments with enchantment runes on item tooltips.")
     public boolean obfuscateCurses = true;
-    @Config(description = "Turn the enchantment glint red when the item is enchanted with at least one curse.")
+    @Config(name = "tint_enchantment_glint_when_cursed", description = "Tint the enchantment glint in a red shade when the item is enchanted with curses.")
     public boolean cursedItemGlint = true;
-    @Config(description = "Hide curses from enchanted books.")
+    @Config(description = "Obfuscate curses on enchanted books.")
     public boolean affectBooks = true;
-    @Config(name = "shift_shows_curses", description = "Temporarily show curses as normal text while any shift key is held in creative mode.")
+    @Config(name = "shift_shows_curses_in_creative", description = "Temporarily show curses as normal text while any shift key is held in creative mode.")
     public boolean shiftShows = true;
-    @Config(description = "Amount of experience levels required to make curses on an item visible by combining with an item from the 'sneakycurses:reveal_curses' tag (amethyst shards by default) in an anvil.")
+    @Config(name = "material_cost_for_revealing_curses_in_anvil", description = "Amount of experience levels required to make curses on an item visible by combining with an item from the 'sneakycurses:reveal_curses' tag (amethyst shards by default) in an anvil.")
     @Config.IntRange(min = 1)
     public int revealCursesCost = 5;
     @Config(description = "Chance wearing or using a cursed piece of equipment will trigger the curses to be revealed. Set to 0.0 to disable revealing curses this way.")