+++ 1.20/Common/src/main/java/fuzs/sneakycurses/data/client/ModLanguageProvider.java
@@ -12,6 +12,6 @@ public class ModLanguageProvider extends AbstractLanguageProvider {
 
     @Override
     protected void addTranslations(TranslationBuilder builder) {
-        builder.add(CurseRevealHandler.KEY_ITEM_CURSES_REVEALED, "Curses revealed for %s");
+        builder.add(CurseRevealHandler.KEY_ITEM_CURSES_REVEALED, "Curses revealed for %s...");
     }
 }