+++ 1.20/CHANGELOG.md
@@ -3,7 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
-## [v8.0.0-1.20.1] - 2023-06-27
-- Ported to Minecraft 1.20.1
+## [v8.0.0-1.20.1] - 2023-10-04
+- Rewritten for Minecraft 1.20.1
+### Changed
+- Curses are no longer hidden from tooltips, instead they are obfuscated in enchantment runes
+- Obfuscated curses can bei revealed by combining the item with an amethyst shard in an anvil
+- Alternatively equipping and using an item will reveal curses on it after some time
+- Cursed item names are no longer colored red, instead the enchantment glint of the item itself shows with a red tint
 
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/